move fidoNameInput file
