+++ packages/core-mobile/app/new/routes/onboarding/seedless/(fido)/_layout.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import { Stack } from 'common/components/Stack'
+
+export default function FidoLayout(): JSX.Element {
+  return (
+    <Stack>
+      <Stack.Screen name="fidoNameInput" />
+    </Stack>
+  )
+}
